refactor directory structure for user and questions
