+++ services/question-service/package.json
@@ -2,7 +2,7 @@
   "name": "question-service",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "ts-node-dev start.ts",
+    "dev": "ts-node-dev src/start.ts",
     "lint:fix": "eslint --ext .ts --fix ."
   },
   "devDependencies": {
