+++ services/question-service/src/domain/question-schema.ts
