+++ services/question-service/src/domain/question-use-case.ts
@@ -1,5 +1,5 @@
 import * as questionRepo from "../data-access/question-repository";
-import type { QuestionFilter, QuestionRequest } from "../types";
+import type { QuestionFilter, QuestionRequest } from "../../types";
 
 import { AddQuestionSchema, UpdateQuestionSchema } from "./question-schema"; // Import your schemas here
 import {
